+++ packages/components/pages/f-checkout/CHANGELOG.md
@@ -3,6 +3,18 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+v4.7.0
+------------------------------
+*October 24, 2022*
+
+### Added
+- Get address from cookies logic
+- Get geolocation from cookies logic
+- Suburb field for address format in New Zealand
+- New prop `shouldLoadAddressFromLocalStorage`
+
+### Removed
+-  Get address from local storage logic in the handler of UPDATE_STATE mutation
 
 v4.6.1
 ------------------------------